refactor: 게임 멤버 도메인, 엔티티 분리
